Move alias tests to new file
